requirements.txt
python-docx>=0.8.10
bibtexparser>=1.4.0
